seperate out explore data toolbar into its own component
